@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Linio\Component\Cache\Adapter;
 
+use Aerospike;
 use Linio\Component\Cache\Exception\InvalidConfigurationException;
 use Linio\Component\Cache\Exception\KeyNotFoundException;
 
@@ -12,20 +13,9 @@ class AerospikeAdapter extends AbstractAdapter implements AdapterInterface
     const BIN_KEY = 'v';
     const DEFAULT_AEROSPIKE_NAMESPACE = 'test';
 
-    /**
-     * @var \Aerospike
-     */
-    protected $db;
-
-    /**
-     * @var string
-     */
-    protected $aerospikeNamespace;
-
-    /**
-     * @var int
-     */
-    protected $ttl;
+    protected Aerospike $db;
+    protected string $aerospikeNamespace;
+    protected int $ttl = 0;
 
     public function __construct(array $config)
     {
@@ -49,7 +39,7 @@ class AerospikeAdapter extends AbstractAdapter implements AdapterInterface
         }
 
         if (isset($config['ttl'])) {
-            $this->ttl = $config['ttl'];
+            $this->ttl = (int) $config['ttl'];
         }
 
         if (isset($config['cache_not_found_keys'])) {
@@ -70,6 +60,9 @@ class AerospikeAdapter extends AbstractAdapter implements AdapterInterface
         }
     }
 
+    /**
+     * @return mixed
+     */
     public function get(string $key)
     {
         $namespacedKey = $this->getNamespacedKey($key);
@@ -109,6 +102,9 @@ class AerospikeAdapter extends AbstractAdapter implements AdapterInterface
         return $values;
     }
 
+    /**
+     * @param mixed $value
+     */
     public function set(string $key, $value): bool
     {
         $status = $this->db->put($this->getNamespacedKey($key), $this->createBin($value), $this->ttl);
@@ -165,17 +161,23 @@ class AerospikeAdapter extends AbstractAdapter implements AdapterInterface
         return true;
     }
 
-    protected function getNamespacedKey(string $key)
+    protected function getNamespacedKey(string $key): array
     {
         return $this->db->initKey($this->aerospikeNamespace, str_replace([':', '.'], '_', $this->namespace), $key);
     }
 
+    /**
+     * @param mixed $value
+     */
     protected function createBin($value): array
     {
         return [self::BIN_KEY => $value];
     }
 
-    protected function removeBin($metadata)
+    /**
+     * @param mixed $metadata
+     */
+    protected function removeBin($metadata): ?string
     {
         if (!is_array($metadata) || !array_key_exists('bins', $metadata) || !is_array($metadata['bins']) || !array_key_exists(static::BIN_KEY, $metadata['bins'])) {
             return null;
