@@ -7,12 +7,14 @@ namespace Linio\Component\Cache\Encoder;
 interface EncoderInterface
 {
     /**
-     * @param $value
+     * @param mixed $value
      */
     public function encode($value): string;
 
     /**
-     * @param $value
+     * @param mixed $value
+     *
+     * @return mixed
      */
     public function decode($value);
 }
